♻️ refactor(mq-lang): extract shared SSRF defenses into resolver::ssrf
is_global_ip, SsrfSafeResolver, and the ureq agent construction were
split across http_import.rs and http_resolver.rs, both only used for
HTTP module imports. Moves them into a dedicated ssrf module so the
upcoming http_get/http_post builtins can reuse the same SSRF-hardened
agent (HTTPS-only, no redirects, DNS results filtered to public IPs)
instead of duplicating this security-sensitive logic.

No behavior change; test_is_global_ip moves with the function it tests.
