+++ crates/mq-lang/src/lib.rs
@@ -90,6 +90,8 @@ pub use lexer::token::{StringSegment, Token, TokenKind};
 pub use module::resolver::http_import;
 #[cfg(feature = "http-import")]
 pub use module::resolver::http_resolver::{HttpFetcher, HttpModuleResolver};
+#[cfg(feature = "http-import")]
+pub use module::resolver::ssrf;
 pub use module::{
     BUILTIN_FILE as BUILTIN_MODULE_FILE, Module, ModuleId, ModuleLoader, STANDARD_MODULES, error::ModuleError,
     resolver::DefaultModuleResolver, resolver::ModuleResolver,
