+++ crates/mq-lang/src/module/resolver/http_resolver.rs
@@ -224,43 +224,6 @@ impl UreqFetcher {
 #[cfg(feature = "http-import-ureq")]
 const MAX_MODULE_SIZE: u64 = 1024 * 1024;
 
-/// DNS resolver that drops any resolved address that isn't publicly routable.
-///
-/// Domain allowlisting only checks the *name* in the URL; without this, an
-/// allowlisted domain could still resolve (at connect time, or via a later
-/// DNS rebind) to a loopback/private/link-local address and reach internal
-/// services. Filtering at the resolver level pins the connection to the
-/// addresses validated here, so a later re-resolution can't smuggle in an
-/// internal address.
-#[cfg(feature = "http-import-ureq")]
-#[derive(Debug, Default)]
-struct SsrfSafeResolver(ureq::unversioned::resolver::DefaultResolver);
-
-#[cfg(feature = "http-import-ureq")]
-impl ureq::unversioned::resolver::Resolver for SsrfSafeResolver {
-    fn resolve(
-        &self,
-        uri: &ureq::http::Uri,
-        config: &ureq::config::Config,
-        timeout: ureq::unversioned::transport::NextTimeout,
-    ) -> Result<ureq::unversioned::resolver::ResolvedSocketAddrs, ureq::Error> {
-        let resolved = self.0.resolve(uri, config, timeout)?;
-
-        let mut safe = self.empty();
-        for addr in resolved.iter() {
-            if super::http_import::is_global_ip(addr.ip()) {
-                safe.push(*addr);
-            }
-        }
-
-        if safe.is_empty() {
-            Err(ureq::Error::HostNotFound)
-        } else {
-            Ok(safe)
-        }
-    }
-}
-
 #[cfg(feature = "http-import-ureq")]
 impl HttpFetcher for UreqFetcher {
     fn fetch(&self, url: &str) -> Result<String, ModuleError> {
@@ -295,16 +258,7 @@ impl HttpFetcher for UreqFetcher {
             return Ok(content);
         }
 
-        let config = ureq::Agent::config_builder()
-            .timeout_global(Some(self.timeout))
-            .https_only(true)
-            .max_redirects(0)
-            .build();
-        let agent = ureq::Agent::with_parts(
-            config,
-            ureq::unversioned::transport::DefaultConnector::default(),
-            SsrfSafeResolver::default(),
-        );
+        let agent = super::ssrf::ssrf_safe_agent(self.timeout, true);
 
         let mut response = agent
             .get(url)
